docs: ajustes  para godoc
